Fixed Issue with tile borders
In tilemaps where the origin lies on border of the map extend, a
calculation is performed where terms can be crossed out. Because
of rounding errors (e.g. x = 160.999999 instead of 161) a map coordinate
could be mapped onto the wrong pixel. As a consequence the pixelBounds
could be too large and a transparent border is added to the tiles.