Fixes loading too many tiles
In tilemaps where the origin lies on border of the map extend,
a calculation is performed where terms can be crossed out.
Because of rounding errors (e.g. x / width = 76.999999999936
instead of 77) a map coordinate could be mapped onto the wrong
tile number. As a consequence the TileCoordinate could be one
off and too many tiles would be loaded.